fix(dashboard): keep review plan types independent of browser runtime
